Add '--arbitrary' feature to rm
When given a set of duplicates with no other duplicates elsewhere, by
default, `psc rm` refuses to delete any of the files, because it can't
figure out which one the user would want to keep. The new '--arbitrary'
flag instructs `psc rm` to arbitrarily choose one of the files and
delete the rest.