Do removal in a single database transaction
Refresh collects a list of files that have disappeared and need to be
removed from the database. Prior to this commit, `db.Remove()` was
called in a loop, which was slow. This patch performs all the removals
in a single database transaction, which is much faster when removing a
large number of files.