Delete fewer DB entries during scan with bounds
Before this patch, scanning a directory deleted all database entries for
that directory. This made sense before adding size bounds, but with the
addition of the '-m' / '-M' flags for min/max size, it makes more sense
to only delete files in the database that match the given size. It makes
sense to keep around previously scanned files that have a size that
falls outside the range we are currently scanning.

This makes it so that we can e.g. scan a directory for files > 1M, and
then go back and scan the ones < 1M, and it won't forget the previously
scanned files. Without this patch, the only way to have this effect
would be to do a complete scan again.