Add `psc hash` command to hash a given file
This is mostly for debugging purposes: it computes the full hash of the
given files, prints them, and persists them to the database. It can be
helpful for forcibly recomputing the short/full hash of a file that has
changed.

This command is purposefully not mentioned in the readme right now
because users probably do not need to use it in the common case.