Switch to tree-based representation in database
The main motivation behind this change is to speed up `db.RemoveDir()`,
which is used by 'scan' as part of incremental scanning, as well as by
'forget'. With the previous flat representation, this operation required
a full table scan to find all files that matched a directory prefix.
Because of the substring-based search, there wasn't an easy fix like
adding an index. The result of this full table scan in `db.RemoveDir()`
was that even when doing a simple operation like incrementally scanning
a tiny directory, it would be unnecessarily slow.

This patch switches from a flat representation of the filesystem, where
each file is represented using its full path, to a tree-based
representation, with files pointing to directories and directories
pointing to their parents. With this new representation, certain
operations like `db.RemoveDir()` can be many times faster than before.
Substring comparisons and full table scans are no longer required.

As a nice side effect, this new representation is more space efficient
because it does not redundantly store all the intermediate directory
names.