Fix infinite loop with rm with no candidates
Prior to this patch, `remove1` would hang if given a set of candidates
where all of the candidate files do not exist anymore. `Lookup` would be
called with an empty string, and that would result in an infinite loop
in `pathToDirectoryId`. This patch fixes the issue by making `remove1`
return early if all of the candidates are gone.